Firmware update (v1.3)
See revision history file

+++ Arduino-sketch/oled_display_lib.h
@@ -87,7 +87,7 @@ void  DrawBox(short w, short h);
 void  SH1106_WriteCommand(uint8_t cmd);
 void  SH1106_WriteData(uint8_t data);
 void  SH1106_Init();
-void  SH1106_SetContrast(unsigned level_pc);  // %
+void  SH1106_SetContrast(uint8_t level_pc);  // %
 void  SH1106_ClearGDRAM();
 void  SH1106_WriteBlock(uint16_t *scnBuf, uint16_t x, uint16_t y, uint16_t w, uint16_t h);
 void  SH1106_Test_Pattern();
